refactor client tsx
